first implementation of Channel objects, using redis and nanomsg
how to use
----------
(application A)
from beacon import channels
c = channels.Channel('toto')
c.value = 42

(application B)
from beacon import channels
c = channels.Channel('toto')
print c.value
42
c.value = 51

(application A)
print c.value
51


how does it work ?
------------------
The first instance of Channel object creates a _Bus object,
containing a nanomsg "bus" socket. This bus socket is connected
to other bus sockets using addresses from a set stored in redis.
Then, the bus socket registers itself in this set.
A thread is started to receive messages from bus sockets found
in redis ; when a new message arrives, it comes in 2 parts: one
is channel name, one is channel value. If channel exists in the
instance, channel value is changed otherwise it is discarded.
The _Bus object also creates a "respondent" nanomsg socket, this
is to be able to reply to a surveyor socket.
When a channel object is created it makes a survey to get the
current channel value - as soon as a reply is received from one
of the respondents, it closes the socket. This is to initialize
the value.
When a channel value is changed, it is sent in the bus in order to
update other channel objects in other instances.
An 'atexit' function cleans the redis database. It doesn't matter
if old addresses are stored in the redis database (in case of kill -9
for example), provided that someone look at it from time to time to
clean if necessary.

Be careful
----------
This is only to move small data between processes.