data/node.py: _create_node takes 'info' keyword argument
This is to create nodes with scan info already filled, to avoid
race conditions with node creation / access to scan_info (when
another process is listening for Redis keyspace notification
events).

node_name is filled when creating the node.

new .new_node property for DataNode ; .new_node returns True when
create=True is used to create the node.