common/session.py: removed _get_global_env_dict function
The environment dictionary is now stored in the session object, and no
more relies on 'globals()' ; if no environment dict is passed to setup
function, an empty dict is created except if BLISS is imported from
the Python interpreter. In this case, the env dict is the one from the
interpreter.
'resetup()' function does not need to receive env_dict, since it uses
the environment dict that was used during setup.