Null writer: fix issue #372
Having no writer at all (writer set to None for Scan object) make BLISS
skipping some calls, in particular writer preparation ; in this case,
".prepare_saving()" on acq masters is *not* called.

In the case of Lima master, it does not set mode to MANUAL, which make
Lima checking directory write access.

This commit creates a Null writer, that does nothing, to be used in
Scans instead of passing None as a writer to fix this issue.