controllers/motor.py: extend lock protection in _initialize_axis
Indeed, _initialize_axis can be called concurrently (when initializing
axes from the same controller from multiple greenlets at the same time),
it was wrong to not include the test for self.__initialized_axis in
the lock.

Changed lock from Semaphore to RLock since the method can be called
in chain.