Add scan.peak(counter, axis=None), scan.com, scan.fwhm functions
Those functions take a counter from the scan, find the corresponding
axis (or use the given one), then uses the peakutils Python module
to do a Gaussian interpolation of the data to find a peak.

scan.peak and scan.com return the axis position of the peak.

If multiple peaks are found, an exception is raised.