writer/file.py: removed code that was causing multiple events received for a master callback
This is a temporary solution until we decide what we really want,
both in term of events for writers and also in the specific case
of HDF5 writer.
At the moment this makes a difference compared to previous implementation,
in case of a timer master slave to a motor master: the timer elapsed_time
will not be repeated both under motor master and timer master (only
within timer master).