data/lima.py: added 'from_stream' property
If 'from_stream' is True, get_image() returns image from video stream
in Lima if the last image is asked for -- this is particularly useful
for displaying (like in Flint).
By default, 'from_stream' is False to make sure latest data is really
returned, from Lima buffer. Take note that the buffer may be empty,
depending on the acquisition.