scan: added support for multiple top-masters
For file writer, and HDF5 in particular, scans with multiple top-masters
are handled as individual scans with a '.X' suffix appended to the
scan number, where X represents the subscan number: 1,2,...

By default any top master will stop the scan when its acquisition is
finished. The 'terminator' property can be set to False to indicate
the scan will not be finished when this top master ends.