Removed Leica microscope controller
This controller was originally written to support the Leica microscope
on ID28 -- the usb communication protocol has been reverse-engineered
to be able to see live video images and to control the microscope
motors.
Later, ID28-specific control was added to be able to do oscillations
(like on MX beamlines), making the Leica controller an aggregate of
devices and turning it into something specific for ID28.