fix issue #630 - TTL setter exception: force redis-py to use SelectSelector
redis-py uses 'poll'-based selector as a default ; 'poll' implementation
in gevent seems to have a problem with redis-py, not sure where it comes
from exactly. However, switching to the 'select'-based selector solves
the problem. As we do not really need the performance improvements
promised by 'poll', it is ok to always use 'select' in redis-py.