shell/cli/repl.py: suppress exception context when raising exception in _execute_task
ptpython tries to execute code with 'eval', if it fails because of SyntaxError it
executes the code with 'exec'. The problem is that the SyntaxError exception
becomes part of the context of any further exception that can be raised.
So, the traceback displays a scary, unhelpful traceback with SyntaxError.
This change removes this problem, by clearing the context.