lakeshore: fixed show method
lakeshore332: subclassing Input,Output,Loop

In loop "show" the letter of the input is used instead of the number.
For writing the curve into the Lakeshore , load is more explicit than write.

Remove *OPC for writing the command.
def send_cmd does not send anymore a"," at the end of the command in case there is no value to write.

It works better wit the channel

Remove *OPC useless and generate ; when reading with the model 335.

Old version without refactoring

According to the Lakeshore model the Input answer is a  number or a letter.

heater range is like the others with a read/set
unit is all models compliant
Heater state of the output is implemented according to temperature.py

Input channel for showing ins compatible with all models. Some of them reply 'A'. Others '1'
_heater_range is like the others with read/set

Status and value of the heater are implemented in the controller because the command is not the same for all the controllers.

It is better to use the method for reading the sensor type

typing error

_heater_range is like the others method with read/set
the state output  inherits from temperature.py the state output

Lazy_init improvement.
For the lakeshore "stop" from the output is not efficient.
I' ll do an issue because it also requires modifications for temperature.py

Add state output from temperature.py
_heater_range becomes with read/set

copy paste error

_heater_range becomes read/set

opu -> ipu because the unity comes frome the input defined in the loop setup in the yml file