issue #847: session: make first setup session the default session
It is allowed to setup multiple sessions, but in the shell the first
one stays the current session: env dict will get populated with
objects from the other sessions, and setup_globals
will contain everything (like nested sessions).
In library mode it is possible to really switch to another session,
but it is left to the user to do it right.