axis: add 'prepare_all' to motor controllers, manage motion hooks in GroupMove
'prepare_move' is renamed to 'get_motion', to not confuse the 'prepare_move'
on motor controller. Now, 'get_motion' does not involve any communication,
the check for readiness of the axis has been moved to GroupMove.

About hooks: pre_move, post_move are called in GroupMove. This allows to
fix issue #1575, hooks are called with the motion list corresponding to the
group move, not for individual moves