fix issue #2411: add transmission counter to CorrCounterController (calc)
Due to the asynchronous nature of calculations performed by calc
controllers, which receive their values from Louie callbacks
from data reading greenlets, calculation inputs have to be complete.
In this issue the CorrCounterController was missing transmission
input => it was read from the filterset itself but, sometimes it
was moving at the moment the Louie callback was processed by
CorrCounterController.
The fix was to simply add 'transmission' counter to inputs and to
deal with the values in calc_function.