pi_e727: fix issue 2608, raw_write_read to not try to read multi line answers ; use raw_write_readlines instead
`raw_write_read` was trying to read multi line answers, and
once a timeout was received it meant 'message complete'.
However, timeout error in comm/tcp has the side effect of
closing the connection ; next command was re-opening a connection.
Sometimes, this reconnection occurs too fast and it is refused
by the controller.
In reality, it is much easier to know which commands return multi line
answers and to wait for those instead (no guessing).